requirements.txt
--extra-index-url https://pypi.anaconda.org/rapidsai-wheels-nightly/simple
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cuda-bindings==12.8.0
cuda-python==12.8.0
cudf-cu12==25.4.0a133
cugraph-cu12==25.4.0a28
cugraph-pyg-cu12==25.4.0a21
cupy-cuda12x==13.3.0
dask-cuda==25.4.0a14
dask-cudf-cu12==25.4.0a133
distributed-ucxx-cu12==0.43.0a10
exceptiongroup==1.2.2
fastrlock==0.8.3
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
libcudf-cu12==25.4.0a133
libcugraph-cu12==25.4.0a28
libkvikio-cu12==25.4.0a11
libraft-cu12==25.4.0a19
libucx-cu12==1.18.0
libucxx-cu12==0.43.0a10
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numba-cuda==0.2.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.10
orjson==3.10.15
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
psutil==6.1.1
pyarrow==19.0.0
Pygments==2.19.1
pylibcudf-cu12==25.4.0a133
pylibcugraph-cu12==25.4.0a28
pylibraft-cu12==25.4.0a19
pylibwholegraph-cu12==25.4.0a21
pynvjitlink-cu12==0.5.0
pynvml==12.0.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pyTigerGraph==1.8.4
pytz==2025.1
PyYAML==6.0.2
raft-dask-cu12==25.4.0a19
rapids-dask-dependency==25.4.0a1
requests==2.32.3
rich==13.9.4
rmm-cu12==25.4.0a16
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tblib==3.0.0
tensordict==0.7.0
threadpoolctl==3.5.0
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ucx-py-cu12==0.43.0a10
ucxx-cu12==0.43.0a10
urllib3==2.3.0
validators==0.34.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
